+++ src/metadata/ExcludeMetadata.ts
@@ -1,5 +1,0 @@
-import { ExcludeOptions } from './ExposeExcludeOptions';
-
-export class ExcludeMetadata {
-  constructor(public target: Function, public propertyName: string, public options: ExcludeOptions) {}
-}