+++ src/metadata/ExposeMetadata.ts
@@ -1,5 +1,0 @@
-import { ExposeOptions } from './ExposeExcludeOptions';
-
-export class ExposeMetadata {
-  constructor(public target: Function, public propertyName: string, public options: ExposeOptions) {}
-}